feat: first entities
