--- backend/src/core/entities/student.ts
@@ -1,1 +1,0 @@
-mock
