+++ backend/src/core/entities/teacher.ts
@@ -1,0 +1,29 @@
+import { TeacherInterface } from "./teacher_interface";
+
+export class Teacher implements TeacherInterface {
+
+    // Necessary variables
+    id: string;     // Teacher id
+    email: string;  // Teacher's email
+    first_name: string;     // Teacher's first name
+    family_name: string;    // Teacher's family name
+    password_hash: string;  // Teacher's hashed password
+
+    // Optional variables
+    name_school?: string;   // Teacher's school
+
+    constructor(
+        id: string, 
+        email: string, 
+        first_name: string, 
+        family_name: string, 
+        password_hash: string, 
+        name_school?: string) {
+        this.id = id;
+        this.email = email;
+        this.first_name = first_name;
+        this.family_name = family_name;
+        this.password_hash = password_hash;
+        this.name_school = name_school;
+    }
+}
